Initial integration of ES to cetera
Initial implementation of an ES client that can speak to a cluster,
using the official ES java client and the transport protocol.

Please make sure you set the cluster name to "catalog" in
/usr/local/opt/elasticsearch/config/elasticsearch.yml
assuming a brew install of ES.  For other installation paths,
the path to the file may be different.