CATALOG-55: Add timing metrics to return body
Added timing metrics both for the service and for ES calls.
Probably going to find out that the the reliance in Streams
for our return types in formatting body, will end in tears
and well, yeah, streams.  But, you live and you learn.
Some of the mess of the code is due to not actually having
objects from ES, but instead just jvalues.  So we have to play
some games to make it work. And it does.