Parameterize datatype boosting and remove default boosts
In the leadup to the summit, we added datatype-specific boosting because
it seemed like the right thing to do (and product said we should). The
issue with this is that we didn't actually measure the impact of
datatype the boosts that we added (10x for datasets and 5x for data
lenses). These boosts actually hurt our relevance more then they
helped. See the results summary
[here](https://gist.github.com/rlvoyer/46587c9c3c1327803e17). This
investigation was prompted by a complaint by one of our new catalog
customers, as described
[in this ticket](https://socrata.atlassian.net/browse/EN-432).

This PR includes the following changes to address this:

- removed default type-specific boosts from reference.conf
- added ability to specify datatype boosts as query params (which
  override defaults)
- more strictly typed defaultTypeBoosts map in ElasticSearchClient
- a few additional unit tests
- added documentation regarding available datatype boosting params
- renamed boosts to fieldBoosts where appropriate

Reviewed-by: ayn.leslie@socrata.com, jason.kroll@socrata.com