Clean up domain filter handling
Previously, we were wrapping the `cname` parameter in an Option to
avoid a possible NullPointerException. In reality, a) we should never
have nulls there, and b) if we were consistently checking for nulls, we
would be doing so unnecessarily and all over the place. Additionally, we
would fail before this
point (https://github.com/socrata/cetera/blob/master/src/main/scala/com/socrata/cetera/util/package.scala#L8-L11)
if null parameters were a possibility. This PR simplifies the
domainFilter code, adding an overloaded domainFilter method for
convenience, and modifying the relevant test.