Move field boosts to should clause
Previously, we were applying our field boosts in the must clause. This
was a subtle regression introduced when we made our simple query a bool
query with both a should and must clause. This PR addresses that
issue. In addition to modifying the unit tests to account for this
change, I validated that this change has a postive impact on
relevance. See the results here:

https://gist.github.com/rlvoyer/5ee5ef42eaea1d6b8587