Do exact match on domain cname in DomainClient.find
Previously, we were doing a phrase search on domain cnames in the
DomainClient.find method. We searched for 2x the number of domains
specified in cnames parameter since a phrase search on an analyzed field
could retrieve more results than we actually wanted. This wasn't enough
to guarantee that we would fetch the desired domain(s). This change
removes the need for that 2x search buffer by replacing the phrase match
query with a terms query on the raw domain cname field.

Reviewed-by: john.krah@socrata.com