EN-6279 update mappings and settings to add users
The service (cetera) and the data provider (Rammstein, Tour-bus) agree
on a schema to share. It would be nice to store that in one file, but
today we keep it up to date manually. Some changes (esp. document)
already in effect but not synced to this repo until now.

Adding a new elasticsearch document type for users, along with
custom analyzers for screen name and email.