Use new owner_id field and update mappings for consistency with Rammstein (#169)
This changeset includes two changes. The primary change is to search on
the owner_id field rather than the deprecated owner field when searching
for docs based on owners. The second change is to use the same document
structure and naming conventions for mapping files as Rammstein to make
it easier to keep the two in sync.

Reviewed-by: john.krah