EN-6387 derived views private (#182)
This change adds an endpoint /asset_selector with identical semantics as the standard catalog, but without visibility filters. This endpoint requires users to authenticate and have a role on the context domain. And it adds a parameter `show_visibility` and result field `metadata.visible_to_anonymous` that is post-calculated in cetera service to show if an anonymous user could view the search result as well as the logged in user.

Reviewed-by: Ayn.Leslie-Cook