EN-11601 Cetera: user search by email should not be case sensitive (#244)
This updates Cetera's user search service to downcase screennames and
email addresses when constructing the corresponding filters. This is
necessary because filters do not honor analyzers defined in the mapping.

There is a corresponding Rammstein PR for the mapping change (updating
the email and screename fields to use the lowercasing keyword analyzer
rather than no analyzer). (NB: the keyword analyzer is effectively a
no-op as it treats an entire field value as a single token.)

Reviewed-by: ayn.leslie